requirements.txt
twine
Pillow
